requirements.txt
# scripts
jinja2
pillow
pyyaml

# workflow
langgraph>=1.1.9
langgraph-checkpoint-sqlite>=2.0.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
anthropic>=0.40.0
fal-client>=0.5.0

# tests
pytest>=8.0.0
